feat(react): render A2uiSurface through the node layer
A2uiSurface now constructs a NodeResolver and renders the resolved tree
with the node-view machinery A2uiNodeSurface carried; A2uiNodeSurface is
deleted, and the react shell sample's ?nodes toggle with it. DeferredChild
recursion over raw definitions remains the fallback for child references
the schema does not classify, so catalogs without REF: markers keep
rendering, and implementations without a view still render through their
self-binding render.

Behavior change: a placeholder upgrade replaces the parent's child
reference, so the parent view re-renders once when a late child arrives;
the id-walking path localized that update to the child. The adapter test
pins the new contract.
