+++ renderers/react/src/v0_9/adapter.tsx
@@ -31,7 +31,7 @@ export interface ReactComponentImplementation extends ComponentApi {
   }>;
   /**
    * Renders this component from a resolved `ComponentNode`. Generated by the
-   * factories below. When absent, `A2uiNodeSurface` falls back to `render`
+   * factories below. When absent, `A2uiSurface` falls back to `render`
    * over the raw definitions.
    */
   view?: React.FC<NodeViewProps>;
@@ -104,7 +104,7 @@ export function createComponentImplementation<Api extends ComponentApi>(
   const NodeView: React.FC<NodeViewProps> = ({node, buildChild}) => {
     const {viewProps, context, viewBuildChild} = useNodeView(node, buildChild);
     if (!context) {
-      throw new Error('A2UI component views render only inside A2uiNodeSurface.');
+      throw new Error('A2UI component views render only inside A2uiSurface.');
     }
     return (
       <MemoizedRender props={viewProps as Props} buildChild={viewBuildChild} context={context} />
@@ -135,7 +135,7 @@ export function createBinderlessComponentImplementation(
     // component binds its own values from the context.
     const {context, viewBuildChild} = useNodeView(node, buildChild);
     if (!context) {
-      throw new Error('A2UI component views render only inside A2uiNodeSurface.');
+      throw new Error('A2UI component views render only inside A2uiSurface.');
     }
     return <RenderComponent context={context} buildChild={viewBuildChild} />;
   };
